backend/requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
channels==3.0.4
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
daphne==3.0.2
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-resized==1.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environ==1.0
fake-useragent==1.5.1
gunicorn==22.0.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.0.0
incremental==22.10.0
lxml==5.2.2
lxml_html_clean==0.1.1
meta_ai_api==1.1.7
packaging==24.1
parse==1.20.2
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyee==11.1.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyppeteer==2.0.0
pyquery==2.0.0
python-dotenv==1.0.1
requests==2.32.3
requests-html==0.10.0
service-identity==24.1.0
setuptools==70.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tqdm==4.66.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
w3lib==2.2.1
websockets==10.4
whitenoise==6.6.0
zipp==3.19.2
zope.interface==6.4.post2
